% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXACTMED.R
\name{EXACTMED}
\alias{EXACTMED}
\title{Exact Mediation Effects Computation}
\usage{
EXACTMED(
  DATA,
  A,
  M,
  Y,
  a1,
  a0,
  M_COV = NULL,
  Y_COV = NULL,
  M_COV_cond = NULL,
  Y_COV_cond = NULL,
  adjusted = TRUE,
  interaction = TRUE,
  Firth = FALSE,
  boot = FALSE,
  nboot = 500,
  bootseed = 1991,
  confcoef = 0.95,
  hvalueM = NULL,
  hvalueY = NULL
)
}
\arguments{
\item{DATA}{a named data frame that includes the outcome, exposure and mediator variables as well as the covariates
to be adjusted for in the model. The exposure can be either binary or continuous. If a covariate is categorical,
it has to be included in the data frame as a factor, character or logical variable.}

\item{A}{the name of the exposure variable.}

\item{M}{the name of the mediator variable.}

\item{Y}{the name of the outcome variable.}

\item{a1}{a value corresponding to the high level of the exposure.}

\item{a0}{a value corresponding to the low level of the exposure.}

\item{M_COV}{a vector containing the names of the adjustment variables (covariates) in the mediator model.}

\item{Y_COV}{a vector containing the names of the adjustment variables (covariates) in the outcome model.}

\item{M_COV_cond}{a named vector (atomic vector or list) containing specific values for some or all
of the adjustment covariates \code{M_COV} in the mediator model. Please consult the package vignette for details.}

\item{Y_COV_cond}{a named vector (atomic vector or list) containing specific values for some or all
of the adjustment covariates \code{Y_COV} in the outcome model. Please consult the package vignette for details.}

\item{adjusted}{a logical variable specifying whether to obtain unadjusted or adjusted estimates.
If \code{adjusted == FALSE}, vectors \code{M_COV} and \code{Y_COV} are ignored by the procedure.}

\item{interaction}{a logical variable specifying whether there is exposure-mediator interaction in the outcome model.}

\item{Firth}{a logical variable specifying whether to compute conventional maximum likelihood estimates
or Firth  penalized estimates in the logistic regression models.}

\item{boot}{a logical value specifying whether the confidence intervals are obtained
by the delta method or by the bootstrap.}

\item{nboot}{If \code{boot == TRUE}, the number of bootstrap replications to obtain the confidence intervals.}

\item{bootseed}{If \code{boot == TRUE}, the value of the initial seed (positive integer) for random number generation.}

\item{confcoef}{a number between 0 and 1 for the confidence coefficient (ex:0.95) for the interval estimates.}

\item{hvalueM}{the value corresponding to the high level of the mediator. If the mediator is already coded
as a numerical binary variable taking 0 or 1 values, then the value of this parameter will be ignored.}

\item{hvalueY}{the value corresponding to the high level of the outcome. If the outcome is already coded
as a numerical binary variable taking 0 or 1 values, then the value of this parameter will be ignored.}
}
\value{
Returns natural direct, indirect and total effect estimates as well as controlled direct effects
    estimates on the OR, RR and RD scales.
}
\description{
The EXACTMED function calculates popular causal mediation effects when the outcome and
    the mediator are binary. More precisely, it returns point and interval estimates for the conditional
    natural direct and indirect effects without making any assumption regarding the rareness or commonness
    of the outcome (hence the term exact). EXACTMED adopts a logistic regression specification for both
    the outcome and mediator in order to compute the exact closed-form effects estimators (see the details in Samoilenko and Lefebvre, 2021).
    For completeness, the EXACTMED function also calculates the conditional controlled direct effects
    at both values of the mediator. Natural and controlled effects estimates are reported in three different
    scales: odds ratio (OR), risk ratio (RR) and risk difference (RD). The interval estimates can be
    obtained either by the delta method or the bootstrap.
}
\details{
By default, EXACTMED reports mediation effects evaluated at the sample-specific mean values of the numerical covariates
    (including the dummy variables created internally by the function to represent the categorical covariates).
    In order to estimate mediation effects at specific values of some covariates (that is, stratum-specific effects)
    the user needs to provide named vectors \code{M_COV_cond} and/or \code{Y_COV_cond} containing those values or levels. The adjustment
    covariates appearing in both \code{M_COV} and \code{Y_COV} (common adjustment covariates) must have the same values; otherwise,
    EXACTMED's execution is aborted and an error message is displayed in the R console.
}
\note{
EXACTMED only works for complete data. Users can apply multiple imputation techniques (e.g., R package \emph{mice})
 or remove observations with any missing values (NA) by specifying \code{data = na.omit(mydataset)}.
}
\examples{
EXACTMED(DATA=datamed, A='X', M='M', Y='Y', a1=1, a0=0, M_COV=c('C1', 'C2'), Y_COV=c('C1', 'C2'))

M_COV_cond <-c(C1 =0.1, C2 =0.4)

Y_COV_cond <-c(C1 =0.1, C2 =0.4)

EXACTMED(DATA=datamed, A='X', M='M', Y='Y', a1=1, a0=0, M_COV=c('C1', 'C2'),
         Y_COV=c('C1', 'C2'), M_COV_cond=M_COV_cond, Y_COV_cond=Y_COV_cond)

C1b <- factor(sample(c("a","b","c"), nrow(datamed), replace =TRUE))

datamed$C1 <- C1b

M_COV_cond <-list(C1='c', C2=0.4)

Y_COV_cond <-list(C1='c', C2=0.4)

EXACTMED(DATA=datamed, A='X', M='M', Y='Y',  a1=1, a0=0, M_COV=c('C1', 'C2'),
Y_COV=c('C1', 'C2'), M_COV_cond=M_COV_cond, Y_COV_cond=Y_COV_cond)
}
\references{
Samoilenko M, Lefebvre G. (2021). Parametric Regression-Based Causal Mediation Analysis of Binary Outcomes
and Binary Mediators: Moving Beyond the Rareness or Commonness of the Outcome. American journal of epidemiology, kwab055. Advance online publication. <doi:10.1093/aje/kwab055>

Samoilenko M, Blais L, Lefebvre G. (2018). Comparing logistic and log-binomial models for causal mediation analyses of binary mediators and rare binary outcomes:
evidence to support cross-checking of mediation results in practice. Observational Studies; 2018(4):193-216.
}
